test run

+++ contracts/ERC20Burn.sol
@@ -1,15 +1,15 @@
+// SPDX-License-Identifier: MIT
 pragma solidity 0.8.19;
+
 import "./ERC20.sol";
 
 // ERC20 token with a burn function
 contract ERC20Burn is ERC20("MyToken", "MT", 18) {
-
-    constructor(){
-        _mint(msg.sender, 10**18);
+    constructor() {
+        _mint(msg.sender, 10 ** 18);
     }
 
-    function burn(uint amount) public{
+    function burn(uint256 amount) public {
         _burn(msg.sender, amount);
     }
-
 }
