+++ contracts/helper.sol
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: MIT
 pragma solidity 0.8.19;
 // From https://github.com/crytic/properties/blob/main/contracts/util/PropertiesHelper.sol
 
@@ -26,14 +27,7 @@ abstract contract PropertiesAsserts {
         if (a != b) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                "!=",
-                bStr,
-                ", reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, "!=", bStr, ", reason: ", reason);
             emit AssertEqFail(string(assertMsg));
             assert(false);
         }
@@ -44,14 +38,7 @@ abstract contract PropertiesAsserts {
         if (a != b) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                "!=",
-                bStr,
-                ", reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, "!=", bStr, ", reason: ", reason);
             emit AssertEqFail(string(assertMsg));
             assert(false);
         }
@@ -62,14 +49,7 @@ abstract contract PropertiesAsserts {
         if (a == b) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                "==",
-                bStr,
-                ", reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, "==", bStr, ", reason: ", reason);
             emit AssertNeqFail(string(assertMsg));
             assert(false);
         }
@@ -80,14 +60,7 @@ abstract contract PropertiesAsserts {
         if (a == b) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                "==",
-                bStr,
-                ", reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, "==", bStr, ", reason: ", reason);
             emit AssertNeqFail(string(assertMsg));
             assert(false);
         }
@@ -98,14 +71,7 @@ abstract contract PropertiesAsserts {
         if (!(a >= b)) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                "<",
-                bStr,
-                " failed, reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, "<", bStr, " failed, reason: ", reason);
             emit AssertGteFail(string(assertMsg));
             assert(false);
         }
@@ -116,14 +82,7 @@ abstract contract PropertiesAsserts {
         if (!(a >= b)) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                "<",
-                bStr,
-                " failed, reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, "<", bStr, " failed, reason: ", reason);
             emit AssertGteFail(string(assertMsg));
             assert(false);
         }
@@ -134,14 +93,7 @@ abstract contract PropertiesAsserts {
         if (!(a > b)) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                "<=",
-                bStr,
-                " failed, reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, "<=", bStr, " failed, reason: ", reason);
             emit AssertGtFail(string(assertMsg));
             assert(false);
         }
@@ -152,14 +104,7 @@ abstract contract PropertiesAsserts {
         if (!(a > b)) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                "<=",
-                bStr,
-                " failed, reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, "<=", bStr, " failed, reason: ", reason);
             emit AssertGtFail(string(assertMsg));
             assert(false);
         }
@@ -170,14 +115,7 @@ abstract contract PropertiesAsserts {
         if (!(a <= b)) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                ">",
-                bStr,
-                " failed, reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, ">", bStr, " failed, reason: ", reason);
             emit AssertLteFail(string(assertMsg));
             assert(false);
         }
@@ -188,14 +126,7 @@ abstract contract PropertiesAsserts {
         if (!(a <= b)) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                ">",
-                bStr,
-                " failed, reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, ">", bStr, " failed, reason: ", reason);
             emit AssertLteFail(string(assertMsg));
             assert(false);
         }
@@ -206,14 +137,7 @@ abstract contract PropertiesAsserts {
         if (!(a < b)) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                ">=",
-                bStr,
-                " failed, reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, ">=", bStr, " failed, reason: ", reason);
             emit AssertLtFail(string(assertMsg));
             assert(false);
         }
@@ -224,35 +148,19 @@ abstract contract PropertiesAsserts {
         if (!(a < b)) {
             string memory aStr = PropertiesLibString.toString(a);
             string memory bStr = PropertiesLibString.toString(b);
-            bytes memory assertMsg = abi.encodePacked(
-                "Invalid: ",
-                aStr,
-                ">=",
-                bStr,
-                " failed, reason: ",
-                reason
-            );
+            bytes memory assertMsg = abi.encodePacked("Invalid: ", aStr, ">=", bStr, " failed, reason: ", reason);
             emit AssertLtFail(string(assertMsg));
             assert(false);
         }
     }
 
     /// @notice Clamps value to be between low and high, both inclusive
-    function clampBetween(
-        uint256 value,
-        uint256 low,
-        uint256 high
-    ) internal returns (uint256) {
+    function clampBetween(uint256 value, uint256 low, uint256 high) internal returns (uint256) {
         if (value < low || value > high) {
-            uint ans = low + (value % (high - low + 1));
+            uint256 ans = low + (value % (high - low + 1));
             string memory valueStr = PropertiesLibString.toString(value);
             string memory ansStr = PropertiesLibString.toString(ans);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                valueStr,
-                " to ",
-                ansStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", valueStr, " to ", ansStr);
             emit LogString(string(message));
             return ans;
         }
@@ -260,24 +168,15 @@ abstract contract PropertiesAsserts {
     }
 
     /// @notice int256 version of clampBetween
-    function clampBetween(
-        int256 value,
-        int256 low,
-        int256 high
-    ) internal returns (int256) {
+    function clampBetween(int256 value, int256 low, int256 high) internal returns (int256) {
         if (value < low || value > high) {
-            int range = high - low + 1;
-            int clamped = (value - low) % (range);
+            int256 range = high - low + 1;
+            int256 clamped = (value - low) % (range);
             if (clamped < 0) clamped += range;
-            int ans = low + clamped;
+            int256 ans = low + clamped;
             string memory valueStr = PropertiesLibString.toString(value);
             string memory ansStr = PropertiesLibString.toString(ans);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                valueStr,
-                " to ",
-                ansStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", valueStr, " to ", ansStr);
             emit LogString(string(message));
             return ans;
         }
@@ -287,20 +186,11 @@ abstract contract PropertiesAsserts {
     /// @notice clamps a to be less than b
     function clampLt(uint256 a, uint256 b) internal returns (uint256) {
         if (!(a < b)) {
-            assertNeq(
-                b,
-                0,
-                "clampLt cannot clamp value a to be less than zero. Check your inputs/assumptions."
-            );
+            assertNeq(b, 0, "clampLt cannot clamp value a to be less than zero. Check your inputs/assumptions.");
             uint256 value = a % b;
             string memory aStr = PropertiesLibString.toString(a);
             string memory valueStr = PropertiesLibString.toString(value);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                aStr,
-                " to ",
-                valueStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", aStr, " to ", valueStr);
             emit LogString(string(message));
             return value;
         }
@@ -313,12 +203,7 @@ abstract contract PropertiesAsserts {
             int256 value = b - 1;
             string memory aStr = PropertiesLibString.toString(a);
             string memory valueStr = PropertiesLibString.toString(value);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                aStr,
-                " to ",
-                valueStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", aStr, " to ", valueStr);
             emit LogString(string(message));
             return value;
         }
@@ -331,12 +216,7 @@ abstract contract PropertiesAsserts {
             uint256 value = a % (b + 1);
             string memory aStr = PropertiesLibString.toString(a);
             string memory valueStr = PropertiesLibString.toString(value);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                aStr,
-                " to ",
-                valueStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", aStr, " to ", valueStr);
             emit LogString(string(message));
             return value;
         }
@@ -349,12 +229,7 @@ abstract contract PropertiesAsserts {
             int256 value = b;
             string memory aStr = PropertiesLibString.toString(a);
             string memory valueStr = PropertiesLibString.toString(value);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                aStr,
-                " to ",
-                valueStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", aStr, " to ", valueStr);
             emit LogString(string(message));
             return value;
         }
@@ -372,12 +247,7 @@ abstract contract PropertiesAsserts {
             uint256 value = b + 1;
             string memory aStr = PropertiesLibString.toString(a);
             string memory valueStr = PropertiesLibString.toString(value);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                aStr,
-                " to ",
-                valueStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", aStr, " to ", valueStr);
             emit LogString(string(message));
             return value;
         } else {
@@ -391,12 +261,7 @@ abstract contract PropertiesAsserts {
             int256 value = b + 1;
             string memory aStr = PropertiesLibString.toString(a);
             string memory valueStr = PropertiesLibString.toString(value);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                aStr,
-                " to ",
-                valueStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", aStr, " to ", valueStr);
             emit LogString(string(message));
             return value;
         } else {
@@ -410,12 +275,7 @@ abstract contract PropertiesAsserts {
             uint256 value = b;
             string memory aStr = PropertiesLibString.toString(a);
             string memory valueStr = PropertiesLibString.toString(value);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                aStr,
-                " to ",
-                valueStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", aStr, " to ", valueStr);
             emit LogString(string(message));
             return value;
         }
@@ -428,12 +288,7 @@ abstract contract PropertiesAsserts {
             int256 value = b;
             string memory aStr = PropertiesLibString.toString(a);
             string memory valueStr = PropertiesLibString.toString(value);
-            bytes memory message = abi.encodePacked(
-                "Clamping value ",
-                aStr,
-                " to ",
-                valueStr
-            );
+            bytes memory message = abi.encodePacked("Clamping value ", aStr, " to ", valueStr);
             emit LogString(string(message));
             return value;
         }
@@ -489,7 +344,7 @@ library PropertiesLibString {
                 // Keep dividing temp until zero.
                 temp := div(temp, 10)
 
-                 // prettier-ignore
+                // prettier-ignore
                 if iszero(temp) { break }
             }
 
@@ -506,10 +361,8 @@ library PropertiesLibString {
 
     function toString(address value) internal pure returns (string memory str) {
         bytes memory s = new bytes(40);
-        for (uint i = 0; i < 20; i++) {
-            bytes1 b = bytes1(
-                uint8(uint(uint160(value)) / (2 ** (8 * (19 - i))))
-            );
+        for (uint256 i = 0; i < 20; i++) {
+            bytes1 b = bytes1(uint8(uint256(uint160(value)) / (2 ** (8 * (19 - i)))));
             bytes1 hi = bytes1(uint8(b) / 16);
             bytes1 lo = bytes1(uint8(b) - 16 * uint8(hi));
             s[2 * i] = char(hi);
